examples/sentiment_analysis/requirements.txt
APScheduler==3.11.0
cachetools==6.1.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.2
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
event_pipeline==2.0.0
fonttools==4.58.5
graphviz==0.20.3
idna==3.10
joblib==1.5.1
kiwisolver==1.4.8
lxml==6.0.0
matplotlib==3.10.3
more-itertools==10.6.0
nltk==3.9.1
numpy==2.3.1
packaging==25.0
pillow==11.3.0
ply==3.11
premailer==3.10.0
pydantic_mini==1.0.4
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
regex==2024.11.6
requests==2.32.4
six==1.17.0
textblob==0.19.0
tqdm==4.67.1
treelib==1.7.0
typing_extensions==4.12.2
tzlocal==5.3.1
urllib3==2.5.0
yagmail==0.15.293
